+++ plugin/src/main/kotlin/de/infolektuell/gradle/typst/extensions/TypstExtension.kt
@@ -9,6 +9,7 @@ abstract class TypstExtension {
   abstract val version: Property<String>
     abstract val creationTimestamp: Property<String>
   abstract val compiler: DirectoryProperty
+  abstract val localPackages: DirectoryProperty
   companion object {
     const val EXTENSION_NAME = "typst"
   }